+++ .gitignore
@@ -4,4 +4,3 @@ temp.txt
 /helpers/firebaseconfig.js
 /helpers/webhooks.js
 /helpers/notifications.js
-google-services_v4.json
