+++ controllers/Customer.js
@@ -1,6 +1,5 @@
 
 const CustomerFunctions = require('../functions/Customer');
-const notification = require('../helpers/notification');
 const response = require('../helpers/response');
 const authFunctions = require('../functions/auth');
 
@@ -508,18 +507,6 @@ const getInvoiceById = async (req, res) => {
     }
 }
 
-const getAllMyNotifications = async (req, res) => {
-    try {
-        let Booking = await notification.getAllMyNotifications(req)
-        if (!Booking) return response.resBadRequest(res, "couldn't find Booking")
-        return response.resSuccessData(res, Booking);
-
-    } catch (error) {
-        console.log(error);
-        return response.resInternalError(res, error)
-    }
-}
-
 
 module.exports = {
     getProfile,
@@ -558,5 +545,4 @@ module.exports = {
     deleteShopReviews,
     updatePassword,
     uplaodAvatar,
-    getAllMyNotifications,
 }
