+++ functions/admin.js
@@ -1039,14 +1039,6 @@ const getSalesSingleShop = async (req) => {
     return response
 };
 
-
-// ----------------------------------------------- Notification  -----------------------------------------------------//
-
-
-
-
-
-
 module.exports = {
     getBusinessbyStatus,
     businessApprove,
