+++ functions/auth.js
@@ -24,25 +24,6 @@ const updateRefreshToken = async (req, token, role) => {
     }
 }
 
-const setDeviceId = async (req, role) => {
-    if (role == "customer") {
-        let Customer = await CustomerModel.findOneAndUpdate({ username: req.body.identifier }, { $set: { deviceId: req.body.deviceId } })
-        return Customer
-    }
-    if (role == "seller") {
-        let seller = await SellerModel.findOneAndUpdate({ username: req.body.identifier }, { $set: { deviceId: req.body.deviceId } })
-        return seller
-    }
-    if (role == "admin") {
-        let admin = await AdminModel.findOneAndUpdate({ username: req.body.identifier, role: req.body.role }, { $set: { deviceId: req.body.deviceId } })
-        return admin
-    }
-    if (role == "agent") {
-        let admin = await AdminModel.findOneAndUpdate({ username: req.body.identifier, role: req.body.role }, { $set: { deviceId: req.body.deviceId } })
-        return admin
-    }
-}
-
 const getUser = async (req, role) => {
 
     if (role == "customer") {
@@ -102,7 +83,7 @@ const isOTPAlreadySended = async (req) => {
     if (OTP.length < 1) return false;
     let createdAt = new Date(OTP[OTP.length - 1]._doc.createdAt)
     let currentDate = new Date
-    let isExpire = (currentDate - createdAt) > 1000 * 60 * 2
+    let isExpire = (currentDate - createdAt) > 1000 * 60  * 2
     console.log(isExpire)
     return !isExpire;
 
@@ -131,6 +112,5 @@ module.exports = {
     resetPassword,
     isOTPAlreadySended,
     getAdminByEmail,
-    getAdmin,
-    setDeviceId,
+    getAdmin
 }
