+++ models/Customer.js
@@ -25,7 +25,6 @@ const CustomerSchema = new Schema({
         id: { type: mongoose.Types.ObjectId },
         role: { type: String, enum: ['admin'] }
     },
-    deviceId: { type: String },
     isActive: { type: Boolean, default: true },
     terminateAt: { type: Date },
     isVerifed: { type: Boolean, default: false },
