+++ models/admin.js
@@ -23,7 +23,6 @@ const AdminSchema = new Schema({
         id: { type: mongoose.Types.ObjectId },
         role: { type: String, enum: ['admin'] }
     },
-    deviceId: { type: String },
     isActive: { type: Boolean, default: true },
     terminateAt: { type: Date },
     isVerifed: { type: Boolean, default: false },
