+++ models/seller.js
@@ -22,7 +22,6 @@ const SellerSchema = new Schema({
         id: { type: mongoose.Types.ObjectId },
         role: { type: String, enum: ['admin'] }
     },
-    deviceId: { type: String },
     isActive: { type: Boolean, default: true },
     terminateAt: { type: Date },
     business: {
